Moved test files from src to tests folder with _test.ts naming
- Moved all *.spec.ts files from src/ to tests/ folder
- Renamed test files from *.spec.ts to *_test.ts format
- Updated all import paths in test files to reference ../src/
- Fixed all jest.mock() paths to match new structure
- Updated tsconfig.json to remove rootDir restriction
- Updated tsconfig.lint.json to include tests folder
- Updated jest.config.js to exclude *_test.ts from coverage
- Updated steering files to use npm instead of pnpm
- All 85 tests passing
- Build and lint successful
